Started on converting single script to modules. Almost nothing works in this version. It is just a beginning.
