+++ index.html
@@ -7,10 +7,10 @@
     <title>SVG Icons v0.1</title>
     <meta name='viewport' content='width=device-width, initial-scale=1'>
     <link rel='stylesheet' type='text/css' media='screen' href='style.css'>
-    <script src='script.js'></script>
+    <script type="module" src='script/app.mjs'></script>
 </head>
 
-<body onload="setup()">
+<body>
     <div class="app">
         <div class="header">
             <h1>SVG Icons v0.1</h1>
